cleaner and faster build
