+++ .drone.yml
@@ -18,13 +18,10 @@ pipeline:
         fi
       - export BUILD_DIR=$PWD
       - mkdir $SBT_VOLATILE_TARGET && mount -t tmpfs -o size=4g tmpfs $SBT_VOLATILE_TARGET ; df -h
-      - sbt ++$SCALA_VERSION! "fmt --test";
       - echo "ensimeServerJars := Nil" > ensime.sbt ;
         echo "ensimeServerProjectJars := Nil" >> ensime.sbt ;
-      - sbt ++$SCALA_VERSION! ensimeConfig "all compile test:compile it:compile doc ensime/assembly";
-      - if [ -n "$DRONE_PULL_REQUEST" ] ; then cd testing/cache && sbt ++$SCALA_VERSION! ensimeConfig ensimeServerIndex && cd ../.. ; fi
-      - if [ -n "$DRONE_PULL_REQUEST" ] ; then sbt ++$SCALA_VERSION! "testOnly -- -l tags.IgnoreOnDrone" ; fi
-      - if [ -n "$DRONE_PULL_REQUEST" ] ; then sbt ++$SCALA_VERSION! "it:testOnly -- -l tags.IgnoreOnDrone" ; fi
+      - sbt ++$SCALA_VERSION! "all check prep cpl doc";
+      - if [ -n "$DRONE_PULL_REQUEST" ] ; then sbt ++$SCALA_VERSION! ";testOnly -- -l tags.IgnoreOnDrone ;it:testOnly -- -l tags.IgnoreOnDrone" ; fi
       - if [ -z "$DRONE_PULL_REQUEST" ] ; then sbt ++$SCALA_VERSION! publish ; fi
 
 matrix:
