+++ testing/simple/src/main/scala/util/badconvention.scala
@@ -1,2 +1,4 @@
 // this file exists for the SourceResolverSpec
 package org.util.set
+
+object BadThingHere
